Changed samples buffer data type to double, to improve accuracy

+++ include/WavFileReader.h
@@ -28,7 +28,7 @@ public:
 
     Status open(const char * path);
     void close();
-    Status read(size_t samplesNum, float * buff, size_t * samplesRead);
+    Status read(size_t samplesNum, double * buff, size_t * samplesRead);
     size_t getSamplesTotal();
     size_t getSamplesRead();
     size_t getSamplesLeft();
